+++ CotEditor/Sources/Collection+String.swift
@@ -38,15 +38,15 @@ extension Collection where Element == String {
         let spaceSuffix = suffix.flatMap { " " + $0 } ?? ""
         
         let (rootName, baseCount): (String, Int?) = {
-            let regex = try! NSRegularExpression(pattern: spaceSuffix + "( ([0-9]+))?$")
+            let suffixPattern = NSRegularExpression.escapedPattern(for: spaceSuffix)
+            let regex = try! NSRegularExpression(pattern: suffixPattern + "( ([0-9]+))?$")
             
             guard let result = regex.firstMatch(in: proposedName, range: proposedName.nsRange) else { return (proposedName, nil) }
             
             let root = (proposedName as NSString).substring(to: result.range.location)
-            
             let numberRange = result.range(at: 2)
             
-            guard numberRange.location != NSNotFound else { return (root, nil) }
+            guard numberRange != .notFound else { return (root, nil) }
             
             let number = Int((proposedName as NSString).substring(with: numberRange))