+++ CotEditor/Sources/LineNumberView.swift
@@ -107,11 +107,11 @@ final class LineNumberView: NSView {
     
     private var numberOfLines = 1
     private var drawingInfo: DrawingInfo?
+    private var opacityObserver: NSObjectProtocol?
     private var textObserver: NSObjectProtocol?
     private var selectionObserver: NSObjectProtocol?
     private var frameObserver: NSObjectProtocol?
     private var scrollObserver: NSObjectProtocol?
-    private var opacityObserver: NSObjectProtocol?
     private var colorObserver: NSKeyValueObservation?
     
     private weak var draggingTimer: Timer?
@@ -140,15 +140,12 @@ final class LineNumberView: NSView {
     // MARK: Lifecycle
     
     deinit {
-        [self.textObserver,
-         self.selectionObserver,
-         self.frameObserver,
-         self.scrollObserver,
-         self.opacityObserver]
-            .compactMap { $0 }
-            .forEach { NotificationCenter.default.removeObserver($0) }
+        self.removeTextViewObservers()
         
-        self.colorObserver?.invalidate()
+        if let observer = self.opacityObserver {
+            NotificationCenter.default.removeObserver(observer)
+        }
+        
         self.draggingTimer?.invalidate()
     }
     
@@ -183,21 +180,35 @@ final class LineNumberView: NSView {
     }
     
     
-    /// observe window opacity change
-    override func viewDidMoveToWindow() {
+    /// receiver is about to be attached to / detached from a window
+    override func viewWillMove(toWindow newWindow: NSWindow?) {
         
-        super.viewDidMoveToWindow()
+        super.viewWillMove(toWindow: newWindow)
+        
+        // remove observations before all observed objects are deallocated
+        if newWindow == nil {
+            self.removeTextViewObservers()
+        }
         
         if let observer = self.opacityObserver {
             NotificationCenter.default.removeObserver(observer)
             self.opacityObserver = nil
         }
+    }
+    
+    
+    /// observe window opacity change
+    override func viewDidMoveToWindow() {
+        
+        super.viewDidMoveToWindow()
         
         // ignore when detached
         guard let window = self.window else { return }
         
         // perform redraw on window opacity change
-        self.opacityObserver = NotificationCenter.default.addObserver(forName: DocumentWindow.didChangeOpacityNotification, object: window, queue: .main) { [unowned self] _ in
+        self.opacityObserver = NotificationCenter.default.addObserver(forName: DocumentWindow.didChangeOpacityNotification, object: window, queue: .main) { [weak self] _ in
+            guard let self = self else { return assertionFailure() }
+            
             self.setNeedsDisplay(self.visibleRect)
         }
     }
@@ -387,8 +398,11 @@ final class LineNumberView: NSView {
     /// observe textView's update to update line number drawing
     private func observeTextView(_ textView: NSTextView) {
         
-        self.textObserver = NotificationCenter.default.addObserver(forName: NSText.didChangeNotification, object: textView, queue: .main) { [unowned self] (notification) in
-            guard let textView = notification.object as? NSTextView else { return assertionFailure() }
+        self.textObserver = NotificationCenter.default.addObserver(forName: NSText.didChangeNotification, object: textView, queue: .main) { [weak self] (notification) in
+            guard
+                let self = self,
+                let textView = notification.object as? NSTextView
+                else { return assertionFailure() }
             
             if self.orientation == .horizontal {
                 // -> Count only if really needed since the line counting is high workload, especially by large document.
@@ -398,21 +412,58 @@ final class LineNumberView: NSView {
             self.needsDisplay = true
         }
         
-        self.selectionObserver = NotificationCenter.default.addObserver(forName: EditorTextView.didLiveChangeSelectionNotification, object: textView, queue: .main) { [unowned self] _ in
-            self.needsDisplay = true
+        self.selectionObserver = NotificationCenter.default.addObserver(forName: EditorTextView.didLiveChangeSelectionNotification, object: textView, queue: .main) { [weak self] _ in
+            self?.needsDisplay = true
         }
         
-        self.frameObserver = NotificationCenter.default.addObserver(forName: NSView.frameDidChangeNotification, object: textView, queue: .main) { [unowned self] _ in
-            self.needsDisplay = true
+        self.frameObserver = NotificationCenter.default.addObserver(forName: NSView.frameDidChangeNotification, object: textView, queue: .main) { [weak self] _ in
+            self?.needsDisplay = true
         }
         
-        self.scrollObserver = NotificationCenter.default.addObserver(forName: NSView.boundsDidChangeNotification, object: textView.enclosingScrollView?.contentView, queue: .main) { [unowned self] _ in
-            self.needsDisplay = true
+        self.scrollObserver = NotificationCenter.default.addObserver(forName: NSView.boundsDidChangeNotification, object: textView.enclosingScrollView?.contentView, queue: .main) { [weak self] _ in
+            self?.needsDisplay = true
         }
         
-        self.colorObserver = textView.observe(\.backgroundColor) { [unowned self] (_, _)  in
-            self.needsDisplay = true
+        self.colorObserver?.invalidate()
+        self.colorObserver = textView.observe(\.backgroundColor) { [weak self] (_, _)  in
+            self?.needsDisplay = true
         }
+    }
+    
+    
+    /// remove observers observing textView
+    private func removeTextViewObservers() {
+        
+        if let observer = self.textObserver {
+            assert(self.textView != nil)
+            
+            NotificationCenter.default.removeObserver(observer)
+            self.textObserver = nil
+        }
+        
+        if let observer = self.selectionObserver {
+            assert(self.textView != nil)
+            
+            NotificationCenter.default.removeObserver(observer)
+            self.selectionObserver = nil
+        }
+        
+        if let observer = self.frameObserver {
+            assert(self.textView != nil)
+            
+            NotificationCenter.default.removeObserver(observer)
+            self.frameObserver = nil
+        }
+        
+        if let observer = self.scrollObserver {
+            assert(self.textView?.enclosingScrollView?.contentView != nil)
+            
+            NotificationCenter.default.removeObserver(observer)
+            self.scrollObserver = nil
+        }
+        
+        self.colorObserver?.invalidate()
+        self.colorObserver = nil
     }
     
 }