+++ CotEditor/Sources/StatusBarController.swift
@@ -76,16 +76,16 @@ final class StatusBarController: NSViewController {
             .showStatusBarLine,
             .showStatusBarColumn,
             ]
-        self.defaultsObservers += UserDefaults.standard.observe(keys: editorDefaultKeys) { [unowned self] (_, _) in
-            self.updateEditorStatus()
+        self.defaultsObservers += UserDefaults.standard.observe(keys: editorDefaultKeys) { [weak self] (_, _) in
+            self?.updateEditorStatus()
         }
         let documentDefaultKeys: [DefaultKeys] = [
             .showStatusBarEncoding,
             .showStatusBarLineEndings,
             .showStatusBarFileSize,
             ]
-        self.defaultsObservers += UserDefaults.standard.observe(keys: documentDefaultKeys) { [unowned self] (_, _) in
-            self.updateDocumentStatus()
+        self.defaultsObservers += UserDefaults.standard.observe(keys: documentDefaultKeys) { [weak self] (_, _) in
+            self?.updateDocumentStatus()
         }
     }
     
@@ -100,7 +100,7 @@ final class StatusBarController: NSViewController {
         self.documentAnalyzer?.needsUpdateStatusEditorInfo = true
         self.documentAnalyzer?.invalidateEditorInfo()
         
-        if #available(macOS 10.15, *) { } else {
+        if NSAppKitVersion.current < .macOS10_15 {
             self.appearanceObserver?.invalidate()
             self.appearanceObserver = self.view.observe(\.effectiveAppearance) { [weak self] (_, _) in
                 self?.updateEditorStatus()
@@ -117,6 +117,7 @@ final class StatusBarController: NSViewController {
         self.documentAnalyzer?.needsUpdateStatusEditorInfo = false
         
         self.appearanceObserver?.invalidate()
+        self.appearanceObserver = nil
     }