+++ Tests/UserDefaultsObservationTests.swift
@@ -1,0 +1,64 @@
+//
+//  UserDefaultsObservationTests.swift
+//
+//  CotEditor
+//  https://coteditor.com
+//
+//  Created by 1024jp on 2019-11-22.
+//
+//  ---------------------------------------------------------------------------
+//
+//  © 2019 1024jp
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//  https://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+import XCTest
+@testable import CotEditor
+
+class UserDefaultsObservationTests: XCTestCase {
+
+    private static let key = DefaultKey<Bool>("TestKey")
+    
+    
+    override class func tearDown() {
+        
+        super.tearDown()
+        
+        UserDefaults.standard.restore(key: Self.key)
+    }
+    
+    
+    func testKeyObservation() {
+        
+        let expectation = self.expectation(description: "UserDefaults observation")
+        
+        UserDefaults.standard[Self.key] = false
+        
+        let observer = UserDefaults.standard.observe(key: Self.key, options: [.new]) { (change) in
+            XCTAssertNil(change.old)
+            XCTAssertTrue(change.new!)
+            XCTAssertEqual(OperationQueue.current, .main)
+            
+            expectation.fulfill()
+        }
+        
+        UserDefaults.standard[Self.key] = true
+        self.wait(for: [expectation], timeout: .zero)
+        // -> Waiting with zero timeout can be failed when the closure is performed not immediately but in another runloop.
+        
+        observer.invalidate()
+        UserDefaults.standard[Self.key] = false
+    }
+
+}