+++ CHANGELOG.md
@@ -2,7 +2,30 @@
 Change Log
 ==========================
 
-3.8.5 (unreleased)
+3.8.6 (387)
+--------------------------
+
+### Improvements
+
+- Change the default shortcut for Format > Font > “Reset to Default” command to `⌘0` and let “Bigger” command accept also `⌘=`.
+- Add shortcut `⌘/` to “Show/Hide Status Bar” command.
+- Add shortcut `⌘⇧T` to “Show/Hide Tab Bar” command.
+- Avoid showing rainbow cursor when canceling the initial syntax highlight.
+- [trivial] Modify the layout of the progress dialog and the regular expression syntax reference.
+- [trivial] Let input fields in the Edit pane accept values without a percent sign.
+
+
+### Fixes
+
+- Fix an issue where creating multiple cursors by rectangular selection failed under macOS 10.15.
+- Fix an issue where the selection highlight color in inactive windows could make text hard to read under macOS 10.14–10.15.
+- Fix an issue where the progress message by find/replacement was not updated when no occurrence found.
+- Fix a possible crash when an invalid color code is input to the color code panel.
+- Improve stability.
+
+
+
+3.8.5 (384)
 --------------------------
 
 ### Improvements